Lag egen fil med redigeringsmodus for Vurdering
